Integrate board view with layout skeleton
Integrated board view with layout skeleton.
