+++ client/src/pages/Navbar.js
@@ -14,6 +14,7 @@ const useStyles = makeStyles((theme) => ({
     alignItems: "center",
     zIndex: 1300,
     backgroundColor: "white",
+    height: "10vh"
   },
 }));
